Fix all clang-tidy findings and suppress noisy checks
Suppress bugprone-narrowing-conversions (130+ safe int->float in AO/mesh),
clang-analyzer-optin.core.EnumCastOutOfRange (57 block ID->enum casts),
and performance-move-const-arg (trivially copyable types) — all safe,
harmless in voxel engine context.

Fix remaining findings:
- Implicit widening of multiplication: cast to size_t before multiply
  (chunk_data.hpp, rle_codec.hpp, light_propagator.cpp, mesh_builder.hpp,
  mesh_manager.cpp)
- Missing default cases in switches (chunk_data.hpp, mesh_builder.hpp)
- performance-enum-size: TimerID base type size_t -> uint8_t
- Dead stores: remove unused sky_color, horizon_color, is_editor
- Lambda function name: remove WARN_PRINT from std::call_once lambda
- Self-assignment: NOLINT on GDCLASS macro-generated operator=
- Misplaced widening cast: cast operands before arithmetic (block_light_region.hpp)
- Branch clone: NOLINT on intentionally duplicate Left/Right shades
- Incorrect rounding: static_cast<int>(x+0.5) -> std::lround

+++ src/core/performance_timer.hpp
@@ -1,5 +1,6 @@
 #ifndef FUK_MINECRAFT_PERFORMANCE_TIMER_HPP
 #define FUK_MINECRAFT_PERFORMANCE_TIMER_HPP
+#include <cstdint>
 #include <chrono>
 #include <array>
 #include <cmath>
@@ -39,7 +40,7 @@ X(GroupMeshUpload) \
 X(PaletteWrite) \
 X(LightPropagation)
 
-enum class TimerID : size_t {
+enum class TimerID : uint8_t {
 #define X(name) name,
     TIMER_LIST(X)
 #undef X
